refactor: rearrange table components into separate files, change dir structure
