+++ packages/web/src/components/SequenceView/SequenceMarkerMutation.tsx
@@ -3,13 +3,19 @@ import React, { SVGProps } from 'react'
 import type { MutationElementWithId } from 'src/algorithms/types'
 import { getNucleotideColor } from 'src/helpers/getNucleotideColor'
 
-export interface MutationViewProps extends SVGProps<SVGRectElement> {
+export interface SequenceMarkerMutationProps extends SVGProps<SVGRectElement> {
   mutation: MutationElementWithId
   pixelsPerBase: number
   width: number
 }
 
-export function MutationView({ mutation, pixelsPerBase, width, onClick, ...rest }: MutationViewProps) {
+export function SequenceMarkerMutation({
+  mutation,
+  pixelsPerBase,
+  width,
+  onClick,
+  ...rest
+}: SequenceMarkerMutationProps) {
   const { id, pos, queryNuc } = mutation
   const fill = getNucleotideColor(queryNuc)
   const x = pos * pixelsPerBase
